requirements.txt
addict===2.4.0
aiofiles===24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio===4.9.0
certifi===2025.4.26
charset-normalizer==3.4.1
click===8.1.8
contourpy==1.3.2
cycler===0.12.1
defusedxml==0.7.1
einops===0.8.1
fastapi===0.115.12
ffmpy===0.5.0
filelock===3.18.0
fonttools==4.57.0
fsspec===2025.3.2
gradio===5.27.1
gradio_client==1.9.1
groovy===0.1.2
h11===0.16.0
httpcore===1.0.9
httpx===0.28.1
huggingface-hub==0.30.2
hydra-core==1.3.2
idna===3.10
imageio===2.37.0
imgaug===0.4.0
iopath===0.1.10
Jinja2===3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl===0.1.2
mpmath===1.3.0
networkx===3.4.2
numpy===2.2.5
omegaconf==2.3.0
opencv-python>=4.11.0
orjson===3.10.16
packaging==25.0
pandas===2.2.3
pillow===11.2.1
pip===25.1
platformdirs==4.3.7
portalocker==3.1.1
pycocotools==2.0.8
pydantic===2.11.3
pydantic_core==2.33.1
pydub===0.25.1
Pygments===2.19.1
pyparsing==3.2.3
python-dateutil>=2.9.0
python-multipart==0.0.20
pytz===2025.2
PyYAML===6.0.2
regex===2024.11.6
requests===2.32.3
rich===14.0.0
ruff===0.11.7
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scipy===1.15.2
semantic-version==2.10.0
setuptools==80.0.0
shapely===2.1.0
shellingham==1.5.4
six===1.17.0
sniffio===1.3.1
starlette==0.46.2
supervision==0.25.1
sympy===1.14.0
tifffile===2025.3.30
timm===1.0.15
tokenizers==0.21.1
tomlkit===0.13.2
torch===2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm===4.67.1
transformers==4.51.3
typer===0.15.2
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata===2025.2
urllib3===2.4.0
uvicorn===0.34.2
websockets==15.0.1
yapf===0.43.0
ultralytics==8.3.120
open3d